+++ src/main/java/io/strimzi/kafka/bridge/tracing/OpenTelemetryHandle.java
@@ -18,9 +18,9 @@ import io.opentelemetry.context.Scope;
 import io.opentelemetry.context.propagation.TextMapGetter;
 import io.opentelemetry.context.propagation.TextMapPropagator;
 import io.opentelemetry.instrumentation.api.instrumenter.messaging.MessageOperation;
-import io.opentelemetry.instrumentation.kafkaclients.TracingProducerInterceptor;
+import io.opentelemetry.instrumentation.kafkaclients.v2_6.TracingProducerInterceptor;
 import io.opentelemetry.sdk.autoconfigure.AutoConfiguredOpenTelemetrySdk;
-import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
+import io.opentelemetry.semconv.SemanticAttributes;
 import io.strimzi.kafka.bridge.config.BridgeConfig;
 import io.vertx.ext.web.RoutingContext;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
@@ -46,10 +46,15 @@ class OpenTelemetryHandle implements TracingHandle {
 
     private Tracer tracer;
 
+    @SuppressWarnings("deprecation") //SemanticAttributes.{HTTP_METHOD, HTTP_URL} are deprecated we continue to use it here for backwards compatibility with existing deployments.
     static void setCommonAttributes(SpanBuilder builder, RoutingContext routingContext) {
         builder.setAttribute(SemanticAttributes.PEER_SERVICE, KAFKA_SERVICE);
-        builder.setAttribute(SemanticAttributes.HTTP_METHOD, routingContext.request().method().name());
-        builder.setAttribute(SemanticAttributes.HTTP_URL, routingContext.request().uri());
+        builder.setAttribute(SemanticAttributes.HTTP_METHOD, routingContext.request().method().name()); // TODO remove as part of https://github.com/strimzi/strimzi-kafka-bridge/issues/875
+        builder.setAttribute(SemanticAttributes.HTTP_REQUEST_METHOD, routingContext.request().method().name());
+        builder.setAttribute(SemanticAttributes.HTTP_URL, routingContext.request().uri()); // TODO remove as part of https://github.com/strimzi/strimzi-kafka-bridge/issues/875
+        builder.setAttribute(SemanticAttributes.URL_SCHEME, routingContext.request().scheme());
+        builder.setAttribute(SemanticAttributes.URL_PATH, routingContext.request().path());
+        builder.setAttribute(SemanticAttributes.URL_QUERY, routingContext.request().query());
     }
 
     @Override
@@ -64,7 +69,8 @@ class OpenTelemetryHandle implements TracingHandle {
 
     @Override
     public void initialize() {
-        System.setProperty("otel.metrics.exporter", "none"); // disable metrics
+        System.setProperty("otel.metrics.exporter", "none"); // disable otel metrics exporter. This has no effect on metrics in the bridge.
+        System.setProperty("otel.logs.exporter", "none"); // disable otel logs exporter. This has no effect on logging in the bridge.
         // TODO: to remove when Vert.x won't be used anymore and the ThreadLocalContextStorage could be used again
         System.setProperty("io.opentelemetry.context.contextStorageProvider", "io.strimzi.kafka.bridge.tracing.BridgeContextStorageProvider");
         AutoConfiguredOpenTelemetrySdk.initialize();
@@ -89,12 +95,14 @@ class OpenTelemetryHandle implements TracingHandle {
         return spanBuilder;
     }
 
+    @SuppressWarnings("deprecation") //SemanticAttributes.{MESSAGING_DESTINATION_NAME,MESSAGING_DESTINATION_KIND} are deprecated we continue to use it here for backwards compatibility with existing deployments.
     @Override
     public <K, V> void handleRecordSpan(ConsumerRecord<K, V> record) {
         String operationName = record.topic() + " " + MessageOperation.RECEIVE.name().toLowerCase(Locale.ROOT);
         SpanBuilder spanBuilder = get().spanBuilder(operationName);
         spanBuilder.setAttribute(SemanticAttributes.MESSAGING_DESTINATION, record.topic());
-        spanBuilder.setAttribute(SemanticAttributes.MESSAGING_DESTINATION_KIND, SemanticAttributes.MessagingDestinationKindValues.TOPIC);
+        spanBuilder.setAttribute(SemanticAttributes.MESSAGING_DESTINATION_NAME, record.topic()); // TODO remove as part of https://github.com/strimzi/strimzi-kafka-bridge/issues/875
+        spanBuilder.setAttribute(SemanticAttributes.MESSAGING_DESTINATION_KIND, SemanticAttributes.MessagingDestinationKindValues.TOPIC); // TODO remove as part of https://github.com/strimzi/strimzi-kafka-bridge/issues/875
         spanBuilder.setAttribute(SemanticAttributes.MESSAGING_SYSTEM, "kafka");
         Context parentContext = propagator().extract(Context.current(), TracingUtil.toHeaders(record), MG);
         if (parentContext != null) {
@@ -177,10 +185,12 @@ class OpenTelemetryHandle implements TracingHandle {
             propagator().inject(Context.current(), routingContext, (rc, key, value) -> rc.response().headers().add(key, value));
         }
 
+        @SuppressWarnings("deprecation") //SemanticAttributes.HTTP_STATUS_CODE is deprecated we continue to use it here for backwards compatibility with existing deployments.
         @Override
         public void finish(int code) {
             try {
-                span.setAttribute(SemanticAttributes.HTTP_STATUS_CODE, code);
+                span.setAttribute(SemanticAttributes.HTTP_STATUS_CODE, code); // TODO remove as part of https://github.com/strimzi/strimzi-kafka-bridge/issues/875
+                span.setAttribute(SemanticAttributes.HTTP_RESPONSE_STATUS_CODE, code);
                 // OK status is fine for all 2xx HTTP status codes
                 span.setStatus(code >= 200 && code < 300 ? StatusCode.OK : StatusCode.ERROR);
                 scope.close();
@@ -189,10 +199,12 @@ class OpenTelemetryHandle implements TracingHandle {
             }
         }
 
+        @SuppressWarnings("deprecation") //SemanticAttributes.HTTP_STATUS_CODE is deprecated we continue to use it here for backwards compatibility with existing deployments.
         @Override
         public void finish(int code, Throwable cause) {
             try {
-                span.setAttribute(SemanticAttributes.HTTP_STATUS_CODE, code);
+                span.setAttribute(SemanticAttributes.HTTP_STATUS_CODE, code); // TODO remove as part of https://github.com/strimzi/strimzi-kafka-bridge/issues/875
+                span.setAttribute(SemanticAttributes.HTTP_RESPONSE_STATUS_CODE, code);
                 span.setStatus(code == HttpResponseStatus.OK.code() ? StatusCode.OK : StatusCode.ERROR);
                 span.recordException(cause);
                 scope.close();